# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MatchGrants API
  description: >
    Plataforma de scraping/extração de avisos de financiamento (Portugal2030, Compete2030, PRR),
    anúncios de contratação pública (base.gov.pt) e matching de empresas por NIF.

    As rotas de `avisos` e `contratações públicas` seguem a mesma forma (mesmo conjunto de ações sobre um
    único recurso): listar, detalhe, editar, servir documento(s), disparar ingestão.
  version: "1.0.0"
servers:
  - url: /
    description: Relativo ao host onde a app está a correr (dev/produção).
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sessionid
      description: >
        Sessão Django (cookie `sessionid`), obtida em POST /users/login/. Algumas rotas são
        públicas (sem sessão); as que exigem papel específico estão marcadas em cada operação.
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    page_size:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      description: Máximo 200.
  schemas:
...
    PaginationEnvelope:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        num_pages: { type: integer }
      required: [total, page, page_size, num_pages]

    # --- avisos ---------------------------------------------------------
...
    PlannedGrant:
      type: object
      properties:
        id: { type: integer }
        plan_id: { type: integer }
        designation: { type: string }
        programme: { type: string }
        expected_start: { type: string, format: date, nullable: true }
        expected_end: { type: string, format: date, nullable: true }
        fund: { type: string }
        budget: { type: number, nullable: true }
        nuts: { type: string }
    PlannedGrantListEnvelope:
      allOf:
        - $ref: "#/components/schemas/PaginationEnvelope"
        - type: object
          properties:
            planned_grants:
              type: array
              items: { $ref: "#/components/schemas/PlannedGrant" }
...
paths:
...
  /planned-grants/:
    get:
      tags: [planned-grants]
      summary: Avisos previstos com abertura a partir do dia atual
      security: [{ sessionAuth: [] }]
      description: "Papel exigido: admin, commercial_grants ou commercial_public."
      parameters:
        - name: order_by
          in: query
          schema:
            type: string
            enum: [start_earliest, start_latest, end_earliest, end_latest,
                   allocation_highest, allocation_lowest]
            default: start_earliest
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/page_size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlannedGrantListEnvelope" }
  /planned-grants/sync/:
    get:
      tags: [planned-grants]
      summary: Sincroniza o Plano Anual a partir do Excel oficial através de Scrape
      description: >
        Aberto (sem sessão) — pensado para automação diária. Encontra automaticamente o .xlsx
        na página do plano anual, descarrega e faz upsert por plan_id (só grava o que mudou).
        Convites são excluídos.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        "502":
          description: Falha ao aceder à página ou não encontrar o Excel
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  error: { type: string }

  # ============================= NEWSLETTER =============================
...